begin module system for frontend pieces
