+++ src/frontend/cached/mod.rs
@@ -1,0 +1,6 @@
+//! Cache-backed logic, currently based on Redis
+pub mod task_report;
+pub mod worker;
+
+pub use task_report::task_report;
+pub use worker::cache_worker;
