+++ src/frontend/mod.rs
@@ -1,0 +1,4 @@
+//! CorTeX Frontend capabilities - reports, cache, captchas and more
+pub mod cached;
+pub mod captcha;
+pub mod params;
